Cont'd refactoring of params package structure
