+++ entropylab/pipeline/api/in_process_param_store.py
@@ -13,10 +13,9 @@ from entropylab.pipeline.api.param_store import (
     Param,
     _ns_to_datetime,
 )
-from entropylab.pipeline.params.persistence.tinydb.persistence import (
-    Persistence,
-    Commit,
-    Metadata,
+from entropylab.pipeline.params.persistence.persistence import Commit, Metadata
+from entropylab.pipeline.params.persistence.tinydb.tinydbpersistence import (
+    TinyDbPersistence,
 )
 
 
@@ -38,7 +37,7 @@ class InProcessParamStore(ParamStore):
         self.__params: Dict[str, Param] = dict()  # where current params are stored
         self.__tags: Dict[str, List[str]] = dict()  # tags that are mapped to keys
         self.__dirty_keys: Set[str] = set()  # updated keys not committed yet
-        self.__persistence = Persistence(path)
+        self.__persistence = TinyDbPersistence(path)
         self.checkout()
         if theirs is not None:
             self.merge(theirs, merge_strategy)
