+++ entropylab/pipeline/api/tests/test_tinydb_persistence.py
@@ -1,11 +1,11 @@
-from entropylab.pipeline.params.persistence.tinydb.persistence import Persistence
+from entropylab.pipeline.params.persistence.tinydb.tinydbpersistence import TinyDbPersistence
 
 
 """ __generate_commit_id() """
 
 
 def test__generate_commit_id():
-    target = Persistence()
+    target = TinyDbPersistence()
     commit_id1 = target._TinyDBPersistence__generate_commit_id()
     commit_id2 = target._TinyDBPersistence__generate_commit_id()
     assert commit_id1 != commit_id2
