+++ entropylab/pipeline/params/persistence/tinydb/tinydbpersistence.py
@@ -3,15 +3,12 @@ from __future__ import annotations
 import contextlib
 import copy
 import hashlib
-import json
 import os
 import string
-import time
-from dataclasses import dataclass
-from datetime import datetime, timedelta
+from datetime import timedelta
 from pathlib import Path
 from random import SystemRandom
-from typing import Optional, Callable, List, Dict, Set
+from typing import Optional, Callable, List, Set
 
 from filelock import FileLock
 from tinydb import TinyDB, Query
@@ -20,8 +17,11 @@ from tinydb.table import Table, Document
 
 from entropylab.logger import logger
 from entropylab.pipeline.api.errors import EntropyError
-from entropylab.pipeline.api.param_store import _ns_to_datetime
-from entropylab.pipeline.params.persistence.persistence import Persistence
+from entropylab.pipeline.params.persistence.persistence import (
+    Persistence,
+    Commit,
+    Metadata,
+)
 from entropylab.pipeline.params.persistence.tinydb.storage import JSONPickleStorage
 
 CURRENT_VERSION = "0.2"
@@ -34,37 +34,7 @@ VERSION_KEY = "version"
 REVISION_KEY = "revision"
 
 
-@dataclass
-class Commit:
-    params: Dict
-    tags: Dict
-    id: Optional[str] = None  # commit_id
-    timestamp: Optional[int] = None  # nanoseconds since epoch
-    label: Optional[str] = None
-
-    def __post_init__(self):
-        self.id = self.id or ""
-        self.timestamp = self.timestamp or time.time_ns()
-        self.label = self.label or None
-        self.params = self.params or {}
-        self.tags = self.tags or {}
-
-
-class Metadata:
-    def __init__(self, d: Dict = None):
-        self.id: str = ""  # commit_id
-        self.timestamp: int = time.time_ns()  # nanoseconds since epoch
-        self.label: Optional[str] = None
-        if d:
-            self.__dict__.update(d)
-
-    def __repr__(self) -> str:
-        d = self.__dict__.copy()
-        d["timestamp"] = _ns_to_datetime(self.timestamp)
-        return f"<Metadata({_dict_to_json(d)})>"
-
-
-class Persistence(Persistence):
+class TinyDbPersistence(Persistence):
     def __init__(self, path: Optional[str] | Optional[Path] = None):
         if path is None:
             self.__is_in_memory_mode = True
@@ -160,7 +130,6 @@ class Persistence(Persistence):
         dirty_keys: Optional[Set[str]] = None,
     ) -> str:
         commit.id = self.__generate_commit_id()
-        commit.timestamp = time.time_ns()  # nanoseconds since epoch
         commit.label = label
         self.__stamp_dirty_params_with_commit(commit, dirty_keys)
         doc = self.__build_document(commit)
@@ -257,17 +226,6 @@ class Persistence(Persistence):
             )
         else:
             return self.__doc_to_commit(doc)
-
-
-def _dict_to_json(d: Dict) -> str:
-    return json.dumps(d, default=_json_dumps_default, sort_keys=True, ensure_ascii=True)
-
-
-def _json_dumps_default(value):
-    if isinstance(value, datetime):
-        return str(value)
-    else:
-        return value.__dict__
 
 
 def _set_version(db: TinyDB | str, version: str, revision: Optional[str] = ""):
