+++ entropylab/pipeline/results_backend/sqlalchemy/alembic/versions/2022-05-19-09-12-01_06140c96c8c4_wrapping_param_store_values.py
@@ -10,8 +10,9 @@ import shutil
 
 from entropylab.logger import logger
 from entropylab.pipeline.api.errors import EntropyError
-from entropylab.pipeline.params.migrations import migrate_param_store_0_1_to_0_2
-from entropylab.pipeline.params.persistence.tinydb.persistence import \
+from entropylab.pipeline.params.persistence.migrations import \
+    migrate_param_store_0_1_to_0_2
+from entropylab.pipeline.params.persistence.tinydb.tinydbpersistence import \
     _set_version
 from entropylab.pipeline.results_backend.sqlalchemy.alembic.alembic_util import (
     AlembicUtil,
